Move TracingInterface from misc to base
