+++ src/neml2/base/TracingInterface.cxx
@@ -27,7 +27,7 @@
 #include <memory>
 #include <thread>
 
-#include "neml2/misc/TracingInterface.h"
+#include "neml2/base/TracingInterface.h"
 #include "neml2/misc/assertions.h"
 
 namespace neml2
