+++ tests/include/SampleEventTracer.h
@@ -24,7 +24,7 @@
 
 #pragma once
 
-#include "neml2/misc/TracingInterface.h"
+#include "neml2/base/TracingInterface.h"
 
 namespace neml2
 {
